+++ docs/changelog.rst
@@ -10,6 +10,8 @@ To update |kitty|, :doc:`follow the instructions <binary>`.
 - Add an option :opt:`command_on_bell` to run an arbitrary command when
   a bell occurs (:iss:`1660`)
 
+- Add a shortcut to toggle maximized window state :sc:`toggle_maximized`
+
 - Add support for the underscore key found in some keyboard layouts
   (:iss:`1639`)